backend/requirements.txt
flask
flask-cors
numpy
lightgbm
torch
torchvision
timm
transformers
pillow
joblib
